image-content/requirements.txt
pyyaml==6.0.1
requests==2.31.0
urllib3==1.26.18